requirements.txt
PyYAML
pigpio
kivy[full]
kivy_garden.drag_n_drop